refactor(types): modularize page definitions
