+++ packages/types/src/index.d.ts
@@ -3,7 +3,7 @@ export * from "./constants";
 export * from "./ImageOrientation";
 export * from "./InventoryItem";
 export * from "./MediaItem";
-export * from "./Page";
+export * from "./page";
 export * from "./Pricing";
 export * from "./Product";
 export * from "./PublishLocation";
