+++ packages/types/src/page/contact-form-with-map.ts
@@ -1,0 +1,12 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface ContactFormWithMapComponent extends PageComponentBase {
+  type: "ContactFormWithMap";
+  mapSrc?: string;
+}
+
+export const contactFormWithMapComponentSchema = baseComponentSchema.extend({
+  type: z.literal("ContactFormWithMap"),
+  mapSrc: z.string().optional(),
+});
