+++ packages/types/src/page/contact-form.ts
@@ -1,0 +1,14 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface ContactFormComponent extends PageComponentBase {
+  type: "ContactForm";
+  action?: string;
+  method?: string;
+}
+
+export const contactFormComponentSchema = baseComponentSchema.extend({
+  type: z.literal("ContactForm"),
+  action: z.string().optional(),
+  method: z.string().optional(),
+});
