+++ packages/types/src/page/custom-html.ts
@@ -1,0 +1,12 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface CustomHtmlComponent extends PageComponentBase {
+  type: "CustomHtml";
+  html?: string;
+}
+
+export const customHtmlComponentSchema = baseComponentSchema.extend({
+  type: z.literal("CustomHtml"),
+  html: z.string().optional(),
+});
