+++ packages/types/src/page/footer.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface FooterComponent extends PageComponentBase {
+  type: "Footer";
+  links?: { label: string; url: string }[];
+  logo?: string;
+}
+
+export const footerComponentSchema = baseComponentSchema.extend({
+  type: z.literal("Footer"),
+  links: z
+    .array(z.object({ label: z.string(), url: z.string() }))
+    .optional(),
+  logo: z.string().optional(),
+});
