+++ packages/types/src/page/image-slider.ts
@@ -1,0 +1,8 @@
+import type { PageComponentBase } from "./base";
+
+export interface ImageSliderComponent extends PageComponentBase {
+  type: "ImageSlider";
+  slides?: { src: string; alt?: string; caption?: string }[];
+  minItems?: number;
+  maxItems?: number;
+}
