+++ packages/types/src/page/image.ts
@@ -1,0 +1,14 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface ImageComponent extends PageComponentBase {
+  type: "Image";
+  src?: string;
+  alt?: string;
+}
+
+export const imageComponentSchema = baseComponentSchema.extend({
+  type: z.literal("Image"),
+  src: z.string().optional(),
+  alt: z.string().optional(),
+});
