+++ packages/types/src/page/map-block.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface MapBlockComponent extends PageComponentBase {
+  type: "MapBlock";
+  lat?: number;
+  lng?: number;
+  zoom?: number;
+}
+
+export const mapBlockComponentSchema = baseComponentSchema.extend({
+  type: z.literal("MapBlock"),
+  lat: z.number().optional(),
+  lng: z.number().optional(),
+  zoom: z.number().optional(),
+});
