+++ packages/types/src/page/popup-modal.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+import { PageComponentBase, baseComponentSchema } from "./base";
+
+export interface PopupModalComponent extends PageComponentBase {
+  type: "PopupModal";
+  trigger?: "load" | "delay" | "exit";
+  delay?: number;
+  content?: string;
+}
+
+export const popupModalComponentSchema = baseComponentSchema.extend({
+  type: z.literal("PopupModal"),
+  trigger: z.enum(["load", "delay", "exit"]).optional(),
+  delay: z.number().int().min(0).optional(),
+  content: z.string().optional(),
+});
